refactor(test): merge polars and pandas
